wish 연관관계 오류

+++ backend/src/main/java/catchroom/backend/repository/RoomRepository.java
@@ -18,7 +18,7 @@ public class RoomRepository {
         em.persist(room);
     }
 
-    public Room findOne(Long roomId){
+    public Room findOne(Integer roomId){
         return em.find(Room.class,roomId);
     }
